requirements.txt
configparser==4.0.2
contextlib2==0.6.0.post1
distlib==0.3.7
Django==1.8.1
django-cors-headers==1.2.2
django-easy-pdf==0.1.0
django-filter==0.9.2
django-rest-framework==0.1.0
djangorestframework==3.2.2
djangorestframework-jwt==1.9.0
filelock==3.2.1
gunicorn==19.6.0
html5lib==1.1
importlib-metadata==2.1.3
importlib-resources==3.3.1
Markdown==2.5.2
olefile==0.44
pathlib2==2.3.7.post1
Pillow==4.0.0
platformdirs==2.0.2
psycopg2==2.8.6
psycopg2-binary==2.7.3.2
PyJWT==1.4.2
pyparsing==2.1.10
PyPDF2==1.28.6
python-dateutil==2.5.3
pytz==2016.7
reportlab==3.5.59
requests==2.11.1
scandir==1.10.0
singledispatch==3.7.0
six==1.10.0
typing==3.10.0.0
#vboxapi==1.0
virtualenv==20.15.1
webencodings==0.5.1
xhtml2pdf==0.2.3
XlsxWriter==2.0.0
xlwt==1.3.0
zipp==1.2.0
